refactor: move status list helpers items to a dedicated module
