+++ src/server/handlers/status_list/utils/etag.rs
